Moved I2c to a directory and implemented the blocking variant. Ready for more variants
